Look through completed hosts as well as incomplete hosts in do_one_select_round. This is necessary for the following reason:
During TCP connect scanning, global pings are sent to hosts in the
completed list. When a response was received, do_one_select_round would
search the enitre incomplete list to find the probe that caused it,
which wasn't there. In the meantime, CPU usage would rise because the
entire incomplete list had to be traversed on every call. As more global
pings were sent, eventually the congestion window would fall below that
number and the scan would grind to a halt without the possibility of
being restarted.