Add a parenthesis to the "Expert" mode of the command constructor wizard to say that it opens up the Profile Editor, because that's all it does.
That's confusing. There's already a way to open the profile editor. So this
"Expert" mode should go away. I'm only keeping it around because I think there
should be a way to create a profile by just typing in a command line and
nothing else, which is what I always thought Expert mode should do.