Remove the code that "catches up" the next scheduled send time. It turns out it's beneficial to let this lag a bit to compensate for delays elsewhere in the code, like the pcap reads. If the scan engine can send enough probes it can catch up itself.
This becomes a problem, though, at higher packet rates, when the next scheduled
send time falls behind faster than it can catch up (the actual number of
packets sent falls farther and farther behind the expected number of packets
sent). This gives the scan engine license to blast tons of packets in a vain
attempt to catch up. While it's sending, incoming packets are dropped, causing
a flood of retransmissions, which in turn leaves even less time to listen for
incoming packets. This means that the actual packet per second can decrease as
the requested packets per second increases, once past a certain threshold. This
will be addressed in a forthcoming commit.