Adding -oP for saving a scan in pcap-format.  This works for raw IP packets as well as Ethernet frames (like ARP pings).  It doesn't work for connect() scanning or anything that uses the Nsock library (DNS resolution, NSE, etc.). Most of the logging is done in tcpip.cc so that raw scanning systems like port scanning, OS detection and Traceroute are all covered.
This is useful for saving a scan without having to run tcpdump/wireshark/etc
at the time of the scan.  But my favorate aspect is that it's all from Nmap's
POV: because Nmap uses pcap's filter language to narrow the packets received,
all of the other crap going through the network at the time of the scan isn't
saved as it would be running a sniffer separately.  Only so much can be done
to narrow it down separately because it would be hard to capture ICMP unreach
messages coming to other hosts, for example.

The pcap file is opened up as EN10MB to allow for saving Ethernet frames.
For IP packets, we fill out a "fake" Ethernet header (just in malloc'd space)
with the Ethertype set to IP, and the addresses zeroed.