In setup.py, make functions path_startswith and path_strip_prefix that semi-intelligently deal with path prefixes.
Use the two functions when creating the uninstaller to remove the
installation root (DESTDIR, given with the --root option) from the paths
of the removed files. This is because the uninstaller is meant to run on
the target host, not the build host. If, for example, the build host
staged this installation in /tmp/staging, the target host uninstaller
would try to remove nonexistent files like /tmp/staging/usr/bin/zenmap.

This has no effect if the installation root is undefined.