Creating this branch for OpenSSL testing on Windows.  Thomas Buchanan did most of the initial work, so big thanks to him!
I've added an OpenSSL directory to mswin32 which holds the necessary
DLLs and a guide for building OpenSSL and getting it in place for Nmap.
Thomas's patch is also in place in this commit with little change.

The two DLLs zipped are just shy of 600kb.

Ideally we'll be able to build Nmap with OpenSSL support on Windows w/o
any adjustment since the DLLs will be in place.  OpenSSL only needs to
be built from source when upgrading to a new version.