This is the first preview of the new Search GUI. Currently only the "freetext" search is working, and searching starts once the search string has more than 3 chars in it (temporary).
Also, fixed an AttributeError bug when the search matches a scan that was opened
into a tab in a previous search. It was caused by referencing ScanNotebookPage's
non-existant attribute "command_execution" - it doesn't exist since the scan
wasn't active in this session, but was opened from the DB.

I have also changed the "Unsaved" prefix that's added to a search result if it
matches an open tab to "[Open]". "Unsaved" is ambiguous, since the scans that
were opened from the DB are also tagged as "Unsaved". They were "Saved" to the
DB, so "Unsaved" is misleading.