Handle the case when local.getpreferredencoding() returns None. I got a crash on the SourceForge bug reporter where this was the case.
The preferred encoding is a best guess as to the encoding used by programs. For
example, in some Chinese-language locales, Nmap will print the time zone with
the Big5 encoding. Zenmap needs to decode that into a Python unicode object
before it is shown on the screen or written to a file, or else it will be
displayed as an unintelligible string of bytes.

I have a feeling a better solution than catching the preferred endocing being
None in two separate places is to have NmapParser represent the output as an
opaque unicode object at all times. I have made a note to check this out. It is
a potentially more far-reaching change than this little fix.