All operators are implemented, except "portstate:", which is about to be discussed further and possibly removed from the feature list.
A "+" sign is used instead of a space for keyword search. For example, if you want to search for "linux 2.6", you enter it as "linux+2.6", since the actual space character is used to delimit operators.

There's also some commented out code dealing with the search delay timer, but since it needs thorough testing, I left it out for now.