Add -w/--wait for specifying a connect timeout in milliseconds, add error checking on -t/--idle-timeout value, and change -t docs to say milliseconds rather than the incorrect seconds.
Traditional netcat has -w for connect and read timeouts (none for writes), but
I like the distinction between the already implemented -t and the new -w.
Thoughts?  It won't be any big deal to use the traditional behavior if it's
preferred.

Also, traditional netcat uses seconds for it's timeouts, while these are in
milliseconds (thanks to nsock for making it easy).  Thoughts on this?  I like
the precision, but the difference between 1 and 1000 for specifying a second
could be confusing.  It's plainly noted in the docs though, so it shouldn't be
a big deal.