SSL listening support (including brokering).  Use --ssl (like with connect) to initiate this.
--ssl-cert and --ssl-key are used to specify the certificate and private key
to use.  These options are definitely subject to change.  Any opinions on how
to improve this are most appreciated.  For example, the same file can be used
to hold both the certificate and key, but it doesn't necessarily have to be
that way, so what's a good option for this?

One thing that I consider somewhat important to note is that when reading in
SSL mode, SSL_read() may be called multiple times in the same loop iteration
without calling select() in-between.  This is because SSL may buffer our input
(e.g. if our read buffer isn't big enough for it all), so calling select() won't
necessarily work.  Successive SSL_read()s should NOT block, and this should not
cause any problems because SSL_pending() is used to determine if we should read
again, and that call tells us how many bytes are being buffered by SSL.  I added
a comment in the code giving a brief explanation of this, but I just felt it
useful to add here as well.

The code can use more testing (it doesn't break for me..) and a little more
error-checking, but I've been working on this a couple of days and feel like
committing it :) This is my first experience using the actual OpenSSL library,
so it's been interesting to say the least.