Use unicode objects pervasively when dealing with Nmap text output.
This changes NmapCommand.get_output to always return a unicode object rather
than a plain str for the stdout output of Nmap. Dealing with the output in this
way is necessary because Nmap can print the time zone in a non-ASCII encoding.
Without decoding it the text is interpreted as nonsense ISO-8859-1 bytes.

Previously the output was decoded before displaying it on the screen and before
writing to a file, and kept in the encoded byte string representation othewise.
Now it's represented as a unicode in all cases. The only way to get the raw
byte output is to read the output file directly (still possible through
NmapCommand.get_output_file). Now all the slightly tricky decoding logic is
isolated in one place.

A side effect of this change is that a crash that occurred when you opened a
scan from a file and clicked "Refresh" no longer occurs.