Merge r3121 from Umit. This fixes the higwidgets header comments to show their true license, which is the LGPL.
------------------------------------------------------------------------
r3121 | boltrix | 2008-07-10 06:40:59 -0600 (Thu, 10 Jul 2008) | 1 line

Discussing with Bartosz, we have detected that the header of higwidgets files containing the license claim are worng. The headers were copied from other Umit files which are GPL, and in reallity, higwidgets is LGPL as has always being described by the COPYING_HIGWIDGETS file inside the trunk. This is not a change of license, but a fix on the headers claimings.
------------------------------------------------------------------------