Use the gtk.TextBuffer methods to retrieve the lines from the Nmap output, rather than grabbing all the text and splitting it manually. This ensures that the lines Zenmap is doing regular expression matching on and the lines inside the text buffer are the same.
Previously it could get confused on Windows because Zenmap was splitting lines
on "\n" but there were "\r\r" sequences in the output. The TextBuffer would
split lines in the middle of the sequence where Zenmap would not, leading to
the two notions of lines getting out of sync. This caused a crash like
/gtk/gtktextbtree.c:3968: byte index off the end of the line