Include Ndiff in the py2exe build on Windows. Zenmap and Ndiff share a Python interpreter and runtime.
This comes with a reorganization of the installation of Zenmap, done to keep
ndiff.exe from being confusingly installed in the "zenmap" directory. Now
zenmap.exe and ndiff.exe are installed in C:\Program Files\Nmap along with
nmap.exe. The py2exe runtime is stored in a subdirectory. This has the side
effect that Ndiff and Zenmap are sure to be in the PATH if Nmap is.

As a byproduct of the py2exe build, the following additional files are stored
in C:\Program Files\Nmap, and not in a subdirectory thereof:
	msvcr71.dll python25.dll
Also Zenmap's share directory is now directly in C:\Program Files\Nmap\share.