Divide the measured response rate by the estimated proportion of responsive ports. This gives a usable estimated rate when scanning firewalled hosts that respond to only a fraction of probes.
For example, when scanning a host with 5 open and 995 filtered ports at 100
probes per second, we measure an average response rate of only 5 / 1000 * 100 =
0.5 responses per second. So if rate limiting kicks in it would be at this
absurdly low rate. So we scaled the measure average by (in this case) 1000 / 5,
giving 100, what the response rate would be if all probes elicited a response.

Rate limiting is not likely to have an effect unless most ports are responsive,
in which case the ratio will be close to 1.0 and behavior will be as with
previous revisions. This revision is just aimed at fixing the behavior on
firewalled hosts. Note that the most common ports are more likely to be
detected at the beginning of a scan, so the estimated responsiveness ratio will
be an overestimate.