Allow sending a timing ping not only when 1.25 s have elapsed since the last ping send or packet receipt, but also when at least 50 probes have been sent since the last ping send or packet receipt. Disable ping_magnifier.
The idea behind this is, instead of artificially inflating the value of timing
pings, we simply send more of them at high packet rates. The choice of 50 is
because that's the value of cc_scale_max, the most any reply can be "worth" in
view of non-responses. The problem with sending timing pings at fixed time
intervals is that at higher scan rates more than 150 probes can be sent between
pings. Yet each ping response can stand for at most 150 non-responses
(cc_scale_max * ping_magnifier = 50 * 3 = 150). Sending pings more often allows
the congestion window to grow more normally, as if every probe were eliciting a
response.

In a few quick tests I found that the scheduler lags a bit and timing pings are
sent every 50-60 probes, so in effect timing pings are still not worth as much
as they should be. I believe they will still be closer to their true worth than
before at packet rates of about 144/s or more (144 = 150 / 1.25 * 50 / 60), and
anyway this should be good enough to indicate whether this course is worth
pursuing.