Fix the ABS macro; change it from #define ABS(x) (((x) >= 0)?(x):(-x)) to #define ABS(x) (((x) >= 0)?(x):-(x))
The old definition was wrong because for example ABS(a - b) would expand to
(((a - b) >= 0)?(a - b):(-a - b)), which is negative when a < b and both are
positive.

There is only one place where this change affects behavior: in
ScanProgressMeter::printStatsIfNecessary in Nmap's timing.cc. That controls
when status updates are printed.