Merge from /nmap-exp/david/ncat-allowdeny.
This fixes several problems with Ncat's --allow and --deny options. The syntax
for those is changed to be the same as Nmap's target specifications (including
host names), with the addition of CIDR netmasks for IPv6 addresses. Allow/deny
specifications are checked for syntax, host list files are loaded, and names
are resolved once at program startup. (The old code never checked
specifications for syntax and loaded host lists files each time an address was
checked.) You can combine several specifications by giving the --allow or
--deny option more than once; in the old code later options overwrote earlier
ones.