Make nsock_read_bytes return after reading the given number of bytes, no matter how low it is. Previously the argument was ignored unless it was greater than NSOCK_READ_CHUNK_SIZE, which is 589823. This was a problem for Ncat when reading from a pipe on stdin; even requesting 0 bytes would return half a meg, far too much to send all at once. Now nsock_read_bytes(0) works just like recv(); it returns whatever is immediately available and fits in a buffer.
nsock_read_bytes isn't used anywhere in the main Nmap or Ncat code, but it is
used in a number of NSE scripts and modules through the nmap.receive_bytes
function. I don't think this will break any of them, but it's possible a script
had a byte limit set too low and never noticed it because it always got back
more data than it requested.