Add code that allows checking for input on stdin on Windows without blocking. Modify fselect to use this code for reads from file descriptor 0 on Windows.
The new code is in a new file, nbase_winunix.c. It works by starting a
background thread that constantly reads from stdin and writes to an internal
pipe. File handles are reassigned to make the read end of the pipe look like
stdin to the rest of the program.