Rework informational and debugging messages to make the -v option more useful. There are new loguser and logdebug functions that handle informational and debugging messages, respectively.
A new global option o.debug is incremented with -vv and higher. At -v,
only very few status messages are printed, such as "Connected to...", as
well as error messages. At -vv minor debugging messages are enabled,
like what command is being executed by --sh-exec. -vvv brings in all
debugging messages, and enables Nsock tracing with a level of 1. Level 1
used to be enabled with -v, and it produced a whole lot of output.