Don't use FD_SET and socket_count_zero to manage the select sets for Nsock connections. Use socket_count_{read,write}_{inc,dec} instead.
There was an assertion in socket_count_zero that there was only one
event outstanding, namely, a connect event. That is because
socket_count_zero would clear the select lists and any events that were
relying on bits being set would be ignore if they were cleared.

This assertion was tripped by Nmap's reverse DNS resolution after
r14215, which forced handle_connect_event (and therefore
socket_count_zero) to be called for all connections, even those that
succeed immediately. Contrary to the assumption of socket_count_zero,
nmap_dns.cc queues a connect event, and read event, and a write event
before the connect event is handled.

Previously this issue was skirted by the fact that handle_connect_result
was not called for UDP "connections" because the connections would
succeed immediately. In fact, this was inadvertently required for
correct functioning; before the assertion was in place,
handle_connect_result would have cleared the connect sets and read and
write events would not have been handled.