When verifying a certificate subject name and there are multiple commonName entries, use only the most specific one. I have defined "most specific" like this:
Wildcard patterns are always less specific than non-wildcard patterns.
If both patterns are wildcard or both are non-wildcard, the one with
more name components is more specific. If two names have the same number
of components, the one that comes later in the certificate is more
specific.

All the tough new verification tests pass now.