Merge r16584 from /nmap. ------------------------------------------------------------------------ r16584 | david | 2010-01-26 18:16:14 -0700 (Tue, 26 Jan 2010) | 7 lines
Use a bigger buffer to read IP addresses from /etc/resolv.conf (to
acommonate IPv6 addresses) and make the sscanf format size match the
buffer size (to avoid smashing the stack). The format string is
constructed dynamically to the size of the buffer with Snprintf. Gunnar
Lindberg reported this problem; discussion starts at
http://seclists.org/nmap-dev/2010/q1/250.

------------------------------------------------------------------------