Fix a preprocessor directive. It was #if HAVE_SSL, when it was protecting something that should have #if HAVE_PCAP.
HAVE_SSL isn't defined anywhere. We use HAVE_OPENSSL. This code was
never compiled, in fact it used an incorrect identifier. The penalty for
not including it was that the event's read buffer was left with a
capacity of 0, causing it to have to be allocated on-demand after the
first actual read. That's not a problem. However, it might have been
relying on the filespace buffer being initialized to 0 somewhere, which
it might not have been. I didn't check to see if it was possible that it
might contain garbage data.