save as XML / render to HTML, improve dashboard and scan listings.
results are now only saved as XML, and parsed to HTML right away using the webxml stylesheet. users are taken directly to this HTML page, although it will be integrated in the regular Insecure chrome; download link will also be available.

to better present results, changes to the structure (HTML) and display (CSS) of most pages has been tweaked. notably, some redundancies have been purged and the various items are displayed in a more consistent manner.

scans and their results are limited to the 10 most recent in the UI, although pagination is coming.