Committing path-mtu.nse:
Performs simple Path MTU Discovery to the target host.

TCP or UDP packets are sent to the host with the DF (don't fragment) bit
set and with varying amounts of data.  If an ICMP Fragmentation Needed
is received, or no reply is received after retransmissions, the amount
of data is lowered and another packet is sent.  This continues until
(assuming no errors occur) a reply from the final host is received,
indicating the packet reached the host without being fragmented.

Not all MTUs are attempted so as to not expend too much time or network
resources.  Currently the relatively short list of MTUs to try contains
the plateau values from Table 7-1 in RFC 1191, "Path MTU Discovery".
Using these values significantly cuts down the MTU search space.  On top
of that, this list is rarely traversed in whole because:
    * the MTU of the outgoing interface is used as a starting point, and
    * we can jump down the list when an intermediate router sending a
      "can't fragment" message includes its next hop MTU (as described
      in RFC 1191 and required by RFC 1812)



I had tested this on Linux, NetBSD and Windows XP and it seemed to work
fine; however, David wasn't getting replies to the probes sent.  I replied
(linked below), and I'm particularly interested in testing using UDP and
against different hosts (on a LAN and other across the internet).  I tested
after he posted, and I don't recall any problems still.  I haven't tested
on all systems again since, but quick tests on Linux still seem to work fine
for me.

http://seclists.org/nmap-dev/2010/q3/254
http://seclists.org/nmap-dev/2010/q3/256