o [NSE] Improved ssh2's kex_init() parameters: all of the algorithm   and language lists can be set using new keys in the "options" table   argument. These all default to the same value used before. Also, the   required "cookie" argument is now replaced by an optional "cookie"   key in the "options" table, defaulting to random bytes as the RFC   says the value should be. [Kris]
Only ssh2's fetch_host_key() uses this function, but I'm working on
a script and noticed the design flaw regarding the cookie arg (scripts
shouldn't be required to always pass this when it's specified that its
value should be random).  The rest was added because the default is
just a subset of algorithms offered across implementations.