Change NmapOptions so that it includes the executable name of the command. There is a new member NmapOptions.executable to read and write it. NmapOptions.parse (and NmapOptions.parse_string) now expect the command to start with an executable name, and NmapOptions.render (and NmapOptions.render_string) now prepend the executable to the command. The default executable is "nmap".
This makes the class better fit the usage pattern of almost every place
it was used. In many places there was awkward code to split the first
word off of a list of words so the remainder could be given to
NmapOptions. The executable name was being tacked on to the front of
commands with string concatenation, which caused quoting problems when
the executable contained spaces, like "C:\\Program Files\\Nmap\\nmap.exe".