This fixes a crash when clicking on the Host Viewer button in the Topology tab of Zenmap (after performing a scan).
TypeError: value is of wrong type for this column

This appeared to be caused by HostViewer.py calling get_info for "number_of_scanned_ports" instead of "number_of_open_ports" (it was renamed in RadialNet.py in r20280). RadialNet doesn't recognise the old name so would return None, which caused the error (perhaps we should have better error checking?).