Scroll the NmapOutputViewer onscreen immediately upon updating.
This stops the scroll bar from freaking out (zooming to the top, then
back down to the bottom) on updates. It was caused by scrolling in an
idle handler instead of doing it immediately. We still do it in an idle
handler too (will be a no-op in most cases) because it's possible that
appending to the text buffer makes a scroll bar appear, which re-wraps
the text, which pushes a line past the bottom, breaking auto-scrolling.