This (big) patch introduces a fundamental change within the architecture of nsock in order to prepare the addition of epoll/poll/kqueue... engines.
Instead of having lists of events, with elements pointing to the corresponding
IODs, we now have a single list of IODs within the nsock pool, and each IOD has
its own lists of events.

Timers, that aren't associated to any IOD, are handled separately. The list of
timers is then still in the nsock pool.

Previous design;
----------------
ev0, 1 and 2 are connect events, respectively for iod0, 1 and 2.
ev3 and 4 are read events, respectively for iod0 and 2.

NSP
 |_[connect]: ev0  -->  ev1  -->  ev2
 |           /           |          \
 |       iod0          iod1         iod2
 |          \                      /
 |_[read]: ev3  --> ev4 ----------+
 |
 |_[write]: ...
 |
 |_[timers]: ...


This design is convenient when using the select() function, as checking if an IOD
is active is O(1). Simply iterating over the events (from 0 to 4 here) and call
FD_ISSET() is enough.

With e/poll(), we iterate over the list of "active" fds. This is therefore not
efficient to retrieve the corresponding events.

The proposed design aims to solve that problem.

Proposed design:
----------------
NSP
 |_[timers]: ...         _[connect]: ev1
 |                      |
 |_[iods]:  iod0  -->   iod1  --> iod2
            |                     |_[connect]: ev2
            |                     |_[read]: ev4
            |_[connect]: ev0
            |_[read]: ev3



This structure should be more flexible and work efficiently with both select()
and e/poll() functions.


---
Unfortunately, as this was initially developed for testing, in a personal
repository, this patch mixes cosmetic fixes with the actual changes... Having
this branch on nmap-exp will prevent such things to happen again.