Important architecture changes.
In order to efficiently use epoll, we have to link the events to the
corresponding IODs (within the IOD structures). My first approach was to do this
by putting small lists of events into the IODs. This resulted in a dramatic
increase of the memory footprint.

This patch makes a compromise between the initial model (event lists are stored
in the nsock pool) and the new needs (IODs keep track of the events they're
associated with). Events are grouped by IOD within the event lists, and each IOD
has pointers to its first event of each kind (connect, read, write, pcap).

This model saves CPU (less allocations) and _lots_ of memory. According to some
quick tests with valgrind, nsock-engines now seems to be even lighter than the
current implementation in term of memory usage! :)


NOTE: The PCAP_BSD_SELECT_HACK mode has been temporarily disabled in this
version but I'll restore it ASAP.