Make FPPacket::getPacketBuffer return a newly allocated buffer every time.
Don't make a cached buffer and then reuse it later. That's tricky and (I
think) unnecessary, and it doesn't notice changes to the packet that
happen after the cache is made (which was noted in the comments).

It now returns a 0-length buffer instead of NULL when there are no
PacketElements.

The pkt_len parameter is now mandatory. It doesn't make sense to get a
buffer without knowing its length. It's possible to get the length with
getLength, but now the caller is forced to at least think about the
length when making the call.