Updated zenmap to not delete ports improperly when adding in newer scans.
When newer scans were added in zenmap it determined what ports to delete by seeing if they were closed in the latest scan. This was overly simplistic since the latest scan may not have actually scanned those ports.

Another error was this it did not check whether the latest scan was scanning the same protocol, tcp vs udp etc...

This patch fixes that by checking tha the latest scan has the port in it, that the port was closed, and that the two ports had the same protocol. Additionally it fixes a bug where old port information objects were updated with new port information even though the information was a different protocol.

This patch is small and if you want to see the code modifications just view the diff. However, the overall modifications are changing the network inventory add host code and adding a method to nmapparser which produces a dictionary of port->list of protocols which can be used to figure which port and protocols were scanned.

References:
http://seclists.org/nmap-dev/2011/q3/480