Treat a "too many matches" as a "no matches" for purposes of reporting.
What "too many matches" man mean, is that is that there is no clear
winner, and several classes have the same middling percentage. In this
case what we really need is for the fingerprint to be submitted. The
"too many matches" dialog doesn't ask the user to submit, but the "no
matches" does.