Do necessary copying to propagate cpe from ServiceProbeMatch to serviceDeductions.
While ServiceProbeMatch has room for an unlimited number of cpe
templates, and serviceDeductions has room for an unlimited number of cpe
matches, the statically allocated information passing in intermediate
layers means it's difficult to pass a variable-sized list upwards. So
instead it is limited to 3 cpes, one each of part a, h, and o.