Add an additional hop-by-hop header at the beginning of the IE2 probe.
OS X won't let us open raw sockets with certain protocol values (which
correspond to the Next Header field in the packets we want to send).
Through testing with -sO, I found these values to be:
	43  routing
	44  fragment
	50  ESP
	51  AH
	60  destination
	108 IPcomp
In some cases, this can be worked around by setting ancillary data
before calling sendmsg (see send_ipv6_ip). But the OS then looks at the
ancillary data and has a chance to reject it. It rejects the routing
header we send because it doesn't contain any source-routing addresses
(sendmsg returns "Invalid argument"). I found the code where the kernel
does this; it's in xnu-1504.15.3/bsd/netinet6/ip6_output.c:2858.

But notably, the list above doesn't include protocol 0, hop-by-hop
options. Once we get a socket of the correct protocol we can stuff
whatever data we want in it, including extension headers that the OS
would reject if it examined them more closely. The extra hop-by-hop
header is a guard to protect the other headers we wish to send.