Introduced the new logging mechanism.
Nsock now has the following macros, for internal use:
  nsock_log_error()
  nsock_log_info()
  nsock_log_debug()
  nsock_log_debug_all()

nsock pools also come with a logging function and a current
loglevel.

As expected, these macros call the logging function if their
level is greater than or equal to the current loglevel.

A logging function receives a nsock_log_rec structure, which
contains the message itself as a normal zero-terminated
string, as well as information about where and when it was
issued.

Nsock provides a default logger, that basically prints things
on stderr. This function can be set externally via
nsock_set_logger()